exo cont

+++ finance_mgmt_server/finance_mgmt_WebAPI/Controllers/ExpenditureController.cs
@@ -1,5 +1,6 @@
 ﻿using finance_mgmt_Repository.DataAccessLayer.Abstraction;
 using finance_mgmt_Repository.Models;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
@@ -7,21 +8,41 @@ namespace finance_mgmt_WebAPI.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
+
     public class ExpenditureController : ControllerBase
     {
         private readonly IExpenditure acc;
+      
         public ExpenditureController(IExpenditure acc)
         {
             this.acc = acc;
         }
         [HttpGet]
+   
         public IActionResult GetAllExpenditure()
         {
             try
             {
-                List<Expenditure> allExpenditures = acc.GetAllExpenses();
-                if (allExpenditures.Count > 0) return Ok(allExpenditures);
-                return Ok();
+                var userIdClaim = User?.Identity?.Name;
+                if (int.TryParse( userIdClaim, out var userId))
+                {
+                    List<Expenditure> allExpenditures = acc.GetAllExpenses().Where(exp => exp.UserId == userId).ToList<Expenditure>();
+                    if (allExpenditures.Count > 0) return Ok(allExpenditures);
+                    return Ok();
+                }
+                else
+                {
+                    return BadRequest("Error occur while identifying claim");
+                }
+                //var headerValue = HttpContext.Request.Headers["Authorization"];
+                //if (headerValue.Count > 0)
+                //{
+                //    var arrOfStrings = headerValue.ToString().Split(' ');
+                //    var token = arrOfStrings[1];
+                //    return Ok(userIdClaim);
+                //}
+                //else return BadRequest();
+              
             }
             catch (Exception ex)
             {
@@ -30,18 +51,89 @@ namespace finance_mgmt_WebAPI.Controllers
         }
         [HttpGet]
         [Route("income")]
+       
         public IActionResult GetAllIncome()
         {
             try
             {
-                List<Expenditure> allExpenditures = acc.GetAllIncome();
 
-                if (allExpenditures.Count > 0)
+                var userIdClaim = User?.Identity?.Name;
+                if (int.TryParse(userIdClaim, out var userId))
                 {
-                    
-                    return Ok(allExpenditures);
+                    List<Expenditure> allExpenditures = acc.GetAllIncome().Where(exp => exp.UserId == userId).ToList<Expenditure>(); ;
+
+                    if (allExpenditures.Count > 0)
+                    {
+
+                        return Ok(allExpenditures);
+                    }
+                    return Ok();
                 }
-                return Ok();
+                else
+                {
+                    return BadRequest("Error occur while identifying claim");
+                }
+               
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
+        }
+        [HttpGet]
+        [Route("incomeTable")]
+        public IActionResult GetAllIncomeTablebyCategory()
+        {
+            try
+            {
+
+                var userIdClaim = User?.Identity?.Name;
+                if (int.TryParse(userIdClaim, out var userId))
+                {
+                    List<IncomeTable> allExpenditures = acc.GetAllIncomeTablebyCategory(userId);
+
+                    if (allExpenditures.Count > 0)
+                    {
+
+                        return Ok(allExpenditures);
+                    }
+                    return Ok();
+                }
+                else
+                {
+                    return BadRequest("Error occur while identifying claim");
+                }
+               
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
+        }
+        [HttpGet]
+        [Route("expenseTable")]
+        public IActionResult GetAllExpenseTablebyCategory()
+        {
+            try
+            {
+
+                var userIdClaim = User?.Identity?.Name;
+                if (int.TryParse(userIdClaim, out var userId))
+                {
+                    List<ExpenseTable> allExpenditures = acc.GetAllExpenseTablebyCategory(userId);
+
+                    if (allExpenditures.Count > 0)
+                    {
+
+                        return Ok(allExpenditures);
+                    }
+                    return Ok();
+                }
+                else
+                {
+                    return BadRequest("Error occur while identifying claim");
+                }
+               
             }
             catch (Exception ex)
             {
@@ -54,9 +146,19 @@ namespace finance_mgmt_WebAPI.Controllers
         {
             try
             {
-                Expenditure allExpenditures = acc.FetchbyId(id);
-                if (allExpenditures is not null) return Ok(allExpenditures);
-                return NotFound();
+
+                var userIdClaim = User?.Identity?.Name;
+                if (int.TryParse(userIdClaim, out var userId))
+                {
+                    Expenditure allExpenditures = acc.FetchbyId(id);
+                    if (allExpenditures is not null) return Ok(allExpenditures);
+                    return NotFound();
+                }
+                else
+                {
+                    return BadRequest("Error occur while identifying claim");
+                }
+               
             }
             catch (Exception ex)
             {
@@ -68,8 +170,20 @@ namespace finance_mgmt_WebAPI.Controllers
         {
             try
             {
-                if (acc.AddExpenses(newacc) > 0) return Ok(new { UserID = newacc.ExpId });
-                return NotFound();
+
+                var userIdClaim = User?.Identity?.Name;
+                if (int.TryParse(userIdClaim, out var userId))
+                {
+                    newacc.UserId = userId;
+                    if (acc.AddExpenses(newacc) > 0) return Ok(new { UserID = newacc.ExpId });
+                    return NotFound();
+                }
+                else
+                {
+                    return BadRequest("Error occur while identifying claim");
+                }
+
+               
             }
             catch (Exception ex)
             {
@@ -82,8 +196,19 @@ namespace finance_mgmt_WebAPI.Controllers
         {
             try
             {
-                if (acc.UpdateExpenses(newacc) > 0) return Ok(new { UserID = newacc.ExpId });
-                return NotFound();
+
+                var userIdClaim = User?.Identity?.Name;
+                if (int.TryParse(userIdClaim, out var userId))
+                {
+                    newacc.UserId = userId;
+                    if (acc.UpdateExpenses(newacc) > 0) return Ok(new { UserID = newacc.ExpId });
+                    return NotFound();
+                }
+                else
+                {
+                    return BadRequest("Error occur while identifying claim");
+                }
+               
             }
             catch (Exception ex)
             {
@@ -97,8 +222,18 @@ namespace finance_mgmt_WebAPI.Controllers
         {
             try
             {
-                if (acc.DeleteExpensesbyId(id) > 0) return Ok(new { UserID = id});
-                return NotFound();
+
+                var userIdClaim = User?.Identity?.Name;
+                if (int.TryParse(userIdClaim, out var userId))
+                {
+                    if (acc.DeleteExpensesbyId(id) > 0) return Ok(new { UserID = id });
+                    return NotFound();
+                }
+                else
+                {
+                    return BadRequest("Error occur while identifying claim");
+                }
+              
             }
             catch (Exception ex)
             {
